requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-taggit==0.22.2
idna==2.6
Pillow==5.1.0
pytz==2018.4
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
